THRIFT-108. python: Fix call to PyDict_Next in fastbinary.c
The position parameter is supposed to be a Py_ssize_t*, not an int*.
This was probably just an oversight in the original patch, and it
was preventing maps from serializing properly on x86-64.


git-svn-id: https://svn.apache.org/repos/asf/incubator/thrift/trunk@702640 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/py/src/protocol/fastbinary.c
@@ -419,7 +419,7 @@ output_val(PyObject* output, PyObject* value, TType type, PyObject* typeargs) {
 
   case T_MAP: {
     PyObject *k, *v;
-    int pos = 0;
+    Py_ssize_t pos = 0;
     Py_ssize_t len;
 
     MapTypeArgs parsedargs;
